Ideas for list of predicates

+++ src/test/java/com/example/lambda/MapTest.java
@@ -17,7 +17,6 @@ public class MapTest {
 		System.out.println("== Calc Old Style ==");
 		int sum = 0;
 		int count = 0;
-
 		for (Person p : pl) {
 			if (p.getAge() >= 23 && p.getAge() <= 65) {
 				sum = sum + p.getAge();
@@ -25,19 +24,22 @@ public class MapTest {
 			}
 		}
 
-		long average = sum / count;
+		double average = (double) sum /  (double) count;
 		System.out.println("Total Ages: " + sum);
 		System.out.println("Average Age: " + average);
 
 		// Get sum of ages
 		System.out.println("\n== Calc New Style ==");
-		long totalAge = pl.stream().filter(search.getCriteria("allPilots"))
-				.mapToInt(p -> p.getAge()).sum();
+		long totalAge = pl.stream()
+		   .filter(search.getCriteria("allPilots"))
+		   .mapToInt(p -> p.getAge())
+		   .sum();
 
 		// Get average of ages
 		OptionalDouble averageAge = pl.parallelStream()
-				.filter(search.getCriteria("allPilots"))
-				.mapToDouble(p -> p.getAge()).average();
+		   .filter(search.getCriteria("allPilots"))
+		   .mapToDouble(p -> p.getAge())
+		   .average();
 
 		System.out.println("Total Ages: " + totalAge);
 		System.out.println("Average Age: " + averageAge.getAsDouble());
